backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
langchain==0.3.14
langchain-google-genai
langchain-community==0.3.14
langgraph==0.2.60
chromadb==0.5.23
sentence-transformers==3.3.1
rank-bm25==0.2.2
duckduckgo-search==7.2.1
googlesearch-python
python-multipart==0.0.20
docx2txt==0.8
PyPDF2==3.0.1
aiofiles==24.1.0
